src/requirements.txt
python-dotenv~=0.19.2
Flask~=2.0.2
Flask-JSON
oauthlib
beautifulsoup4
numpy~=1.19.5
pip==21.3.1
wheel==0.37.1
cryptography==36.0.1
lxml==4.8.0
keyring==23.4.1
colorama==0.4.4
pytz==2021.3
Cython~=0.29.28
setuptools==59.6.0
click~=8.0.3
future~=0.18.2
requests==2.27.1
alembic==1.7.6
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
Flask-Bcrypt==0.7.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
pipenv==2022.1.8
pycparser==2.21
python-dateutil==2.8.2
python-editor==1.0.4
python-slugify==6.1.1
six==1.16.0
SQLAlchemy==1.4.31
text-unidecode==1.3
virtualenv==20.13.2
virtualenv-clone==0.5.7
Werkzeug==2.0.3
WTForms==3.0.0
psycopg2==2.7.7